pyproject.toml
[project]
name = "obsidian-llm-wiki"
version = "0.5.0"
description = "Convert Obsidian notes into an AI-maintained wiki using local or cloud LLMs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["obsidian", "wiki", "llm", "ollama", "knowledge-management", "local-ai"]
authors = [{ name = "kytmanov" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "watchdog>=4.0",
    "python-frontmatter>=1.1",
]

[project.urls]
Homepage = "https://github.com/kytmanov/obsidian-llm-wiki-local"
Repository = "https://github.com/kytmanov/obsidian-llm-wiki-local"
Issues = "https://github.com/kytmanov/obsidian-llm-wiki-local/issues"


[project.scripts]
obsidian-llm-wiki = "obsidian_llm_wiki.cli:cli"
olw = "obsidian_llm_wiki.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_llm_wiki"]

[dependency-groups]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "pytest-benchmark>=4.0", "ruff>=0.4"]

[tool.uv]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/obsidian_llm_wiki/providers.py" = ["E501"]  # tabular registry data

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"
markers = ["integration: requires running Ollama instance"]
